+++ src/main/java/com/backend/naildp/controller/HomeController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.backend.naildp.dto.home.PostSummaryResponse;
 import com.backend.naildp.exception.ApiResponse;
+import com.backend.naildp.service.PostInfoService;
 import com.backend.naildp.service.PostService;
 
 import lombok.RequiredArgsConstructor;
@@ -21,7 +22,8 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class HomeController {
 
-	private final PostService postService;
+	// private final PostService postService;
+	private final PostInfoService postInfoService;
 
 	@GetMapping("/home")
 	public ResponseEntity<?> homePosts(
@@ -31,11 +33,11 @@ public class HomeController {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
 		if (authentication == null) {
-			PostSummaryResponse postSummaryResponse = postService.homePosts(choice, size, cursorPostId, "");
+			PostSummaryResponse postSummaryResponse = postInfoService.homePosts(choice, size, cursorPostId, "");
 			return ResponseEntity.ok(ApiResponse.successResponse(postSummaryResponse, "최신 게시물 조회", 2000));
 		}
 
-		PostSummaryResponse postSummaryResponse = postService.homePosts(choice, size, cursorPostId,
+		PostSummaryResponse postSummaryResponse = postInfoService.homePosts(choice, size, cursorPostId,
 			authentication.getName());
 		return ResponseEntity.ok(ApiResponse.successResponse(postSummaryResponse, "최신 게시물 조회", 2000));
 	}
@@ -44,7 +46,7 @@ public class HomeController {
 	public ResponseEntity<?> likedPost(@AuthenticationPrincipal UserDetails userDetails,
 		@RequestParam(required = false, defaultValue = "20", value = "size") int size,
 		@RequestParam(required = false, defaultValue = "-1", value = "oldestPostLikeId") long postLikeId) {
-		PostSummaryResponse likedPostsResponses = postService.findLikedPost(userDetails.getUsername(), size,
+		PostSummaryResponse likedPostsResponses = postInfoService.findLikedPost(userDetails.getUsername(), size,
 			postLikeId);
 		return ResponseEntity.ok(ApiResponse.successResponse(likedPostsResponses, "좋아요 체크한 게시물 조회", 2000));
 	}
